pyproject.toml
[project]
name = "deduplicator-fastapi"
version = "0.1.0"
description = "Deduplication service for product events"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref>=3.8.1",
    "celery>=5.5.1",
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.7",
    "pydantic-settings>=2.8.1",
    "redis>=5.2.1",
]

[dependency-groups]
celery = [
    "celery>=5.5.1",
    "flower>=2.0.1",
]
dev = [
    "black>=25.1.0",
    "locust>=2.34.1",
    "pytest>=8.3.5",
]
